// src/services/model/evaCategorias.ts
/**
 * PostgREST API
 * standard public schema
 *
 * OpenAPI spec version: 7.0.0 (2b61a63)
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface EvaCategorias { 
    /**
     * Note: This is a Primary Key.<pk/>
     */
    catcodigoexterno: string;
    /**
     * Note: This is a Primary Key.<pk/> This is a Foreign Key to `eva_examenes.idexamen`.<fk table='eva_examenes' column='idexamen'/>
     */
    idexamen: number;
    nombre?: string;
    descripcion?: string;
    /**
     * Note: This is a Foreign Key to `eva_estados.idestado`.<fk table='eva_estados' column='idestado'/>
     */
    idestado?: number;
    fecharegistro?: string;
    maxpreguntas?: number;
    minpreguntas?: number;
    aleatorias?: boolean;
}